Completed the program mostly
Got really fustrated and stayed up until 5:30 in the morning banging my head
over a simple school program, but alas I got it done. Decided to use an array list
because i was getting very strange iterator bugs, and eventually realized it was because
of my heavy usage of function calls.

Decided to go with single dimension Integer array lists, for very easy management,
and place the Slidepuzzle in it's own class.